SPEECH/requirements.txt

deep-translator==1.4.4

en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl

gensim==3.8.3
gitdb==4.0.7
GitPython==3.1.18
googletrans==3.0.0

helper==2.5.0





matplotlib==3.4.2


moviepy==1.0.3

nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8

nltk==3.6.2

numpy==1.19.4
opencv-python==4.5.2.54
packaging==20.8
pandas==1.2.4



Pillow==8.2.0
pipreqs==0.4.10
pipwin==0.5.1
plotly==5.1.0
plotly-express==0.4.1



python-dotenv==0.18.0
pyttsx3==2.90



scikit-learn==0.24.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.1

sklearn==0.0


spacy==3.0.6
spacy-legacy==3.0.8
spacy-streamlit==1.0.0
speech==0.5.2
SpeechRecognition==3.8.1

streamlit==0.84.0


sumy==0.8.1

translate==3.6.1

virtualenv==20.4.4

wavio==0.0.4

Werkzeug==1.0.1

xgboost==1.4.2

